+++ docs/changes.md
@@ -3,6 +3,8 @@
 2021/09/10
   * 完善demo\_ui\_ex\_app(感谢兆坤提供补丁)。
   * mledit控件增加追加文本、输入模式（超出最大行数后移除旧的行）功能(感谢兆坤提供补丁)。
+  * fscript\_object 中新增了 object\_forward\_events 接口，并完善了单元测试和说明文档（感谢雨欣提供补丁）
+  * object\_array 添加事件触发，用于mvvm更新界面，并且在 set\_prop 和 get\_prop 函数中增加路径处理，向下查找一层，以便 mvvm 列表渲染逐层访问属性（感谢雨欣提供补丁）
 
 2021/09/09
   * fscript增加json\_save\_to\_string（感谢雨欣提供补丁）